34151: LA port of 8330685: ZGC: share barrier spilling logic
