+++ src/hotspot/cpu/loongarch/gc/shared/barrierSetAssembler_loongarch.cpp
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2018, Oracle and/or its affiliates. All rights reserved.
- * Copyright (c) 2018, 2023, Loongson Technology. All rights reserved.
+ * Copyright (c) 2018, 2024, Loongson Technology. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -34,6 +34,9 @@
 #include "runtime/jniHandles.hpp"
 #include "runtime/sharedRuntime.hpp"
 #include "runtime/stubRoutines.hpp"
+#ifdef COMPILER2
+#include "gc/shared/c2/barrierSetC2.hpp"
+#endif // COMPILER2
 
 #define __ masm->
 
@@ -451,3 +454,78 @@ void BarrierSetAssembler::check_oop(MacroAssembler* masm, Register obj, Register
   __ load_klass(obj, obj); // get klass
   __ beqz(obj, error);     // if klass is null it is broken
 }
+
+#ifdef COMPILER2
+
+OptoReg::Name BarrierSetAssembler::refine_register(const Node* node, OptoReg::Name opto_reg) {
+  if (!OptoReg::is_reg(opto_reg)) {
+    return OptoReg::Bad;
+  }
+
+  const VMReg vm_reg = OptoReg::as_VMReg(opto_reg);
+  if (vm_reg->is_FloatRegister()) {
+    return opto_reg & ~1;
+  }
+
+  return opto_reg;
+}
+
+#undef __
+#define __ _masm->
+
+void SaveLiveRegisters::initialize(BarrierStubC2* stub) {
+  // Record registers that needs to be saved/restored
+  RegMaskIterator rmi(stub->live());
+  while (rmi.has_next()) {
+    const OptoReg::Name opto_reg = rmi.next();
+    if (OptoReg::is_reg(opto_reg)) {
+      const VMReg vm_reg = OptoReg::as_VMReg(opto_reg);
+      if (vm_reg->is_Register()) {
+        _gp_regs += RegSet::of(vm_reg->as_Register());
+      } else if (vm_reg->is_FloatRegister()) {
+        if (UseLASX && vm_reg->next(7))
+          _lasx_vp_regs += FloatRegSet::of(vm_reg->as_FloatRegister());
+        else if (UseLSX && vm_reg->next(3))
+          _lsx_vp_regs += FloatRegSet::of(vm_reg->as_FloatRegister());
+        else
+          _fp_regs += FloatRegSet::of(vm_reg->as_FloatRegister());
+      } else {
+        fatal("Unknown register type");
+      }
+    }
+  }
+
+  // Remove C-ABI SOE registers, scratch regs and _ref register that will be updated
+  if (stub->result() != noreg) {
+    _gp_regs -= RegSet::range(S0, S7) + RegSet::of(SP, SCR1, SCR2, stub->result());
+  } else {
+    _gp_regs -= RegSet::range(S0, S7) + RegSet::of(SP, SCR1, SCR2);
+  }
+}
+
+SaveLiveRegisters::SaveLiveRegisters(MacroAssembler* masm, BarrierStubC2* stub)
+  : _masm(masm),
+    _gp_regs(),
+    _fp_regs(),
+    _lsx_vp_regs(),
+    _lasx_vp_regs() {
+
+  // Figure out what registers to save/restore
+  initialize(stub);
+
+  // Save registers
+  __ push(_gp_regs);
+  __ push_fpu(_fp_regs);
+  __ push_vp(_lsx_vp_regs  /* UseLSX  */);
+  __ push_vp(_lasx_vp_regs /* UseLASX */);
+}
+
+SaveLiveRegisters::~SaveLiveRegisters() {
+  // Restore registers
+  __ pop_vp(_lasx_vp_regs /* UseLASX */);
+  __ pop_vp(_lsx_vp_regs  /* UseLSX  */);
+  __ pop_fpu(_fp_regs);
+  __ pop(_gp_regs);
+}
+
+#endif // COMPILER2
